vfs: introduce VFS_BINDMOUNT_SHIFT_UIDGID
VFS_BINDMOUNT_SHIFT_UIDGID is to Allow mounting the same mountpoint
multiple times but using different UID and GID mappings to each bind mount.
For example, this allows mounting /dir in a user-namespaced container with
the same UID and GID mappings applied as the user-namespace itself.

Introduce the new virtual "vuid_t" and "vgid_t" types and update inode
struct fields accordingly so we can later inspect filesystems and what they
are doing with these values. This allows to catch errors or something we
have overlooked.

The intended behaviour is to do the shift at the VFS layer only,
filesystems should always use the global UID and GID when writing on disk,
which means the UID and GID must be mapped into the global user namespace
before writing. We have introduced the raw_i_uid_read() and raw_i_gid_read()
functions to note that. That's why the vuid_t type is the same of kuid_t
one, this way we are sure that filesystems who still use the i_uid_read()
function at the lower layers will not use the virtual UID and GID that a
process sees inside its user namespace, instead they will continue to use
the global UID and GID to write the correct IDs on disk.

As said the UID and GID shift can be done at the VFS layer when bind
mounting a mountpoint. We pin the user namespace of the caller and use it
to do the mapping. Filesystems can use the new helpers vfs_i_uid_read(),
vfs_i_uid_write() etc to access UIDs and GIDs of an inode. Those
functions use the vfsmount context to do the UID and GID mapping.

Cc: David Herrmann <dh.herrmann@gmail.com>
Cc: Dongsu Park <dongsu@endocode.com>
Signed-off-by: Djalal Harouni <djalal@endocode.com>