vfs:uidshift: pin the userns during a shift bind-mount
Pin the userns during a shift bind-mount and while we are it add the
CL_BIND_SHIFT_UIDGID flag to have more control on the bind mount
operations. As an example unprivileged callers that try to move mounts
between namespaces or any other private mounts are not shifted,
currently they are silently discarded.

Signed-off-by: Djalal Harouni <djalal@endocode.com>