uidshift: introduce MNT_BIND_SHIFTABLE and MS_BIND_SHIFT_UIDGID
Callers can set the MS_BIND_SHIFT_UIDGID flag on a mounted rootfs
when they are real root users. The flag will be propagated to children
mounts and MNT_BIND_SHIFTABLE will be set for each of them.

If the caller wants to bind shift a mount point then at the moment:
1) Pass the "vfs_uidshift=1" option during a bind mount
2) He must be privileged in his own mnt namespace
3) The MNT_BIND_SHIFTABLE flag was previously set on the current
   rootfs.

Signed-off-by: Djalal Harouni <djalal@endocode.com>